Remove run-internal-command file and fix specs
