--- node/src/commands/run-internal-command.js
@@ -1,20 +1,0 @@
-const help = require('./help');
-const configure = require('./configure');
-const pkg = require('../../package.json');
-const logger = require('../lib/logger');
-
-const showVersion = () => {
-  logger.info(pkg.version);
-};
-
-const runInternalCommand = async (command, options) => {
-  const commands = {
-    help: () => help(),
-    version: () => showVersion(),
-    configure: async () => configure(options)
-  };
-
-  await commands[command]();
-};
-
-module.exports = runInternalCommand;
